requirements.txt
pytest==7.3.1
lupa==1.14.1
pyautogui==0.9.54
pynput==1.7.6
pillow==10.2.0
inputs==0.5
setuptools>=65.5.1
wheel>=0.40.0
dxcam==0.0.5